Move dependency downloading and links fixing to html_processor.py and add docstrings
