+++ openedx2zim/xblocks_extractor/html.py
@@ -22,7 +22,7 @@ class Html(BaseXblock):
         html_content = soup.find("div", attrs={"class": "edx-notes-wrapper"})
         if not html_content:
             html_content = str(soup.find("div", attrs={"class": "course-wrapper"}))
-        self.html = self.scraper.dl_dependencies(
+        self.html = self.scraper.html_processor.dl_dependencies_and_fix_links(
             content=html_content,
             output_path=self.output_path,
             path_from_html=self.folder_name,
